wps7.0适配

+++ app/src/main/java/com/nandi/yngsagp/fragment/DangerReportFragment.java
@@ -6,7 +6,7 @@ import android.app.Activity;
 import android.app.AlertDialog;
 import android.app.Fragment;
 import android.app.ProgressDialog;
-import android.content.Context;
+import android.content.ActivityNotFoundException;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.database.Cursor;
@@ -36,6 +36,7 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.view.animation.LinearInterpolator;
 import android.widget.AdapterView;
+import android.widget.ArrayAdapter;
 import android.widget.Button;
 import android.widget.CheckBox;
 import android.widget.Chronometer;
@@ -45,7 +46,6 @@ import android.widget.ImageView;
 import android.widget.LinearLayout;
 import android.widget.PopupWindow;
 import android.widget.RelativeLayout;
-import android.widget.SeekBar;
 import android.widget.Spinner;
 import android.widget.TextView;
 
@@ -63,6 +63,7 @@ import com.nandi.yngsagp.R;
 import com.nandi.yngsagp.adapter.PictureAdapter;
 import com.nandi.yngsagp.bean.AudioPath;
 import com.nandi.yngsagp.bean.DangerUBean;
+import com.nandi.yngsagp.bean.DisasterPoint;
 import com.nandi.yngsagp.bean.PhotoPath;
 import com.nandi.yngsagp.bean.VideoPath;
 import com.nandi.yngsagp.greendao.GreedDaoHelper;
@@ -84,6 +85,7 @@ import java.util.ArrayList;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
+import java.util.Locale;
 import java.util.Map;
 
 import butterknife.BindView;
@@ -101,6 +103,7 @@ public class DangerReportFragment extends Fragment {
     private static final int PICK_PHOTO = 1;
     private static final int TAKE_PHOTO = 2;
     private static final int TAKE_VIDEO = 3;
+    private static final int FILE_SELECT_CODE = 4;
     Unbinder unbinder;
     @BindView(R.id.tab_layout)
     TabLayout tabLayout;
@@ -120,8 +123,6 @@ public class DangerReportFragment extends Fragment {
     TextView latDanger;
     @BindView(R.id.typeDanger)
     Spinner typeDanger;
-    @BindView(R.id.factorDanger)
-    EditText factorDanger;
     @BindView(R.id.personDanger)
     EditText personDanger;
     @BindView(R.id.houseDanger)
@@ -158,6 +159,16 @@ public class DangerReportFragment extends Fragment {
     EditText dReportName;
     @BindView(R.id.ll_dReport)
     LinearLayout llDReport;
+    @BindView(R.id.iv_check_file)
+    Button ivCheckFile;
+    @BindView(R.id.sp_factor_type)
+    Spinner spFactorType;
+    @BindView(R.id.tv_file)
+    TextView tvFile;
+    @BindView(R.id.sp_disaster_point)
+    Spinner spDisasterPoint;
+    @BindView(R.id.ll_upload_file)
+    LinearLayout llUploadFile;
     private int typePos = 0;
     private Activity context;
     private PopupWindow popupWindow;
@@ -172,6 +183,10 @@ public class DangerReportFragment extends Fragment {
     private RequestCall build;
     private MediaPlayer player;
     private String sessionId;
+    private String type;
+    private List<DisasterPoint> disasterPoints;
+    private List<String> disasterName = new ArrayList<>();
+
     @Nullable
     @Override
     public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
@@ -186,7 +201,12 @@ public class DangerReportFragment extends Fragment {
     }
 
     private void initData() {
-        sessionId= (String) SharedUtils.getShare(context,Constant.SESSION_ID,"");
+        disasterPoints = GreedDaoHelper.getDisasterPointList();
+        for (DisasterPoint disasterPoint : disasterPoints) {
+            disasterName.add(disasterPoint.getDisName());
+        }
+        type = (String) SharedUtils.getShare(context, Constant.PERSON_TYPE, "");
+        sessionId = (String) SharedUtils.getShare(context, Constant.SESSION_ID, "");
         DangerUBean queryDanger = GreedDaoHelper.queryDanger();
         List<PhotoPath> queryPhoto = GreedDaoHelper.queryPhoto(2);
         VideoPath queryVideo = GreedDaoHelper.queryVideo(2);
@@ -201,7 +221,7 @@ public class DangerReportFragment extends Fragment {
             latDanger.setText(queryDanger.getLat());
             typePos = Integer.parseInt(queryDanger.getType());
             typeDanger.setSelection(Integer.parseInt(queryDanger.getType()));
-            factorDanger.setText(queryDanger.getFactor());
+            spFactorType.setSelection(Integer.parseInt(queryDanger.getFactor()));
             personDanger.setText(queryDanger.getPerson());
             houseDanger.setText(queryDanger.getHouse());
             moneyDanger.setText(queryDanger.getMoney());
@@ -209,6 +229,7 @@ public class DangerReportFragment extends Fragment {
             otherDanger.setText(queryDanger.getOther());
             dReportMobile.setText(queryDanger.getMobile());
             dReportName.setText(queryDanger.getName());
+            spDisasterPoint.setSelection(queryDanger.getDisasterPosition());
         }
         if (queryPhoto != null && queryPhoto.size() > 0) {
             photoPaths.clear();
@@ -266,6 +287,9 @@ public class DangerReportFragment extends Fragment {
     }
 
     private void setAdapter() {
+        ArrayAdapter<String> arrayAdapter = new ArrayAdapter<String>(context, android.R.layout.simple_list_item_1, disasterName);
+        arrayAdapter.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item);
+        spDisasterPoint.setAdapter(arrayAdapter);
         pictureAdapter = new PictureAdapter(context, photoPaths);
         rvPhoto.setLayoutManager(new GridLayoutManager(context, 3));
         rvPhoto.setAdapter(pictureAdapter);
@@ -377,6 +401,8 @@ public class DangerReportFragment extends Fragment {
         progressDialog.setMessage("正在上传...");
         if ("2".equals(SharedUtils.getShare(context, Constant.PERSON_TYPE, "0"))) {
             llDReport.setVisibility(View.VISIBLE);
+            ivCheckFile.setVisibility(View.VISIBLE);
+            llUploadFile.setVisibility(View.VISIBLE);
         }
         tabLayout.addTab(tabLayout.newTab().setText("文本信息"), 0, true);
         tabLayout.addTab(tabLayout.newTab().setText("媒体信息"), 1);
@@ -396,7 +422,7 @@ public class DangerReportFragment extends Fragment {
         return format.format(date);
     }
 
-    @OnClick({R.id.tv_get_location, R.id.timeDanger, R.id.btn_save, R.id.btn_upload, R.id.iv_take_photo, R.id.iv_take_video, R.id.iv_take_audio, R.id.tv_video, R.id.tv_audio})
+    @OnClick({R.id.tv_get_location, R.id.timeDanger, R.id.btn_save, R.id.btn_upload, R.id.iv_take_photo, R.id.iv_take_video, R.id.iv_take_audio, R.id.tv_video, R.id.tv_audio, R.id.tv_file, R.id.iv_check_file})
     public void onViewClicked(View view) {
         switch (view.getId()) {
             case R.id.timeDanger:
@@ -446,6 +472,26 @@ public class DangerReportFragment extends Fragment {
             case R.id.tv_get_location:
                 startLocation();
                 break;
+            case R.id.tv_file:
+                if (!TextUtils.isEmpty(tvFile.getText())) {
+                    clickText(3);
+                }
+                break;
+            case R.id.iv_check_file:
+                checkFile();
+                break;
+        }
+    }
+
+    private void checkFile() {
+        Intent intent = new Intent(Intent.ACTION_GET_CONTENT);
+        intent.setType("*/*");//过滤文件类型（所有）
+        intent.addCategory(Intent.CATEGORY_OPENABLE);
+
+        try {
+            startActivityForResult(Intent.createChooser(intent, "请选择文件！"), FILE_SELECT_CODE);
+        } catch (ActivityNotFoundException ex) {
+            ToastUtils.showShort("未找到文件管理器");
         }
     }
 
@@ -487,9 +533,10 @@ public class DangerReportFragment extends Fragment {
         String location = locationDanger.getText().toString().trim();
         String lon = lonDanger.getText().toString().trim();
         String lat = latDanger.getText().toString().trim();
-        String type = typePos + "";
-        String factor = factorDanger.getText().toString().trim();
+        String type = typePos - 1 + "";
+        String factor = spFactorType.getSelectedItemPosition() + "";
         String person = personDanger.getText().toString().trim();
+        String disasterId = disasterPoints.get(spDisasterPoint.getSelectedItemPosition()).getDisId();
         if (person.equals(null)) {
             person = "0";
         }
@@ -528,12 +575,13 @@ public class DangerReportFragment extends Fragment {
         map.put("monitorPhone", name);
         map.put("areaId", areaId);
         map.put("personType", personType);
+        map.put("disastersId", disasterId);
         setRequest(map);
     }
 
     private void setRequest(Map<String, String> map) {
         progressDialog.show();
-        PostFormBuilder formBuilder = OkHttpUtils.post().url(getString(R.string.local_base_url) + "appDangerous/add").addHeader("sessionID",sessionId);
+        PostFormBuilder formBuilder = OkHttpUtils.post().url(getString(R.string.local_base_url) + "appDangerous/add").addHeader("sessionID", sessionId);
         for (PhotoPath photoPath : photoPaths) {
             if (photoPath != null) {
                 formBuilder.addFile("file", new SimpleDateFormat("yyyyMMddHHmmss").format(new Date()) + ".jpg", new File(photoPath.getPath()));
@@ -546,7 +594,12 @@ public class DangerReportFragment extends Fragment {
         }
         if (!TextUtils.isEmpty(tvAudio.getText().toString())) {
             Log.d("cp", "音频添加");
-            formBuilder.addFile("file", new SimpleDateFormat("yyyyMMddHHmmss").format(new Date()) + ".mp3", new File(audioPath));
+            formBuilder.addFile("file", new SimpleDateFormat("yyyyMMddHHmmss").format(new Date()) + ".ogg", new File(audioPath));
+        }
+        if (!TextUtils.isEmpty(tvFile.getText().toString())) {
+            File file = new File(tvFile.getText().toString());
+            Log.d("chenpeng", file.getName());
+            formBuilder.addFile("file", file.getName(), file);
         }
         formBuilder.params(map);
         formBuilder.addParams("type", "2");
@@ -575,6 +628,10 @@ public class DangerReportFragment extends Fragment {
                     if (success) {
                         ToastUtils.showShort(data);
                         clean();
+                        if ("2".equals(type)) {
+                            Intent intent = new Intent(Constant.ACTION_DANGER_UPLOADED);
+                            context.sendBroadcast(intent);
+                        }
                     } else {
                         if ("exit".equals(message)) {
                             AppUtils.startLogin(context);
@@ -596,7 +653,7 @@ public class DangerReportFragment extends Fragment {
         latDanger.setText("");
         typePos = 0;
         typeDanger.setSelection(0);
-        factorDanger.setText("");
+        spFactorType.setSelection(0);
         personDanger.setText("");
         houseDanger.setText("");
         moneyDanger.setText("");
@@ -606,6 +663,7 @@ public class DangerReportFragment extends Fragment {
         dReportName.setText("");
         tvAudio.setText("");
         tvVideo.setText("");
+        tvFile.setText("");
         photoPaths.clear();
         pictureAdapter.notifyDataSetChanged();
         DangerUBean dangerUBean = GreedDaoHelper.queryDanger();
@@ -633,6 +691,7 @@ public class DangerReportFragment extends Fragment {
 
     private void save() {
         DangerUBean dangerUBean = new DangerUBean();
+        int disasterPosition = spDisasterPoint.getSelectedItemPosition();
         String reportMan = userDanger.getText().toString().trim();
         String phone = phoneDanger.getText().toString().trim();
         String time = timeDanger.getText().toString().trim();
@@ -641,7 +700,7 @@ public class DangerReportFragment extends Fragment {
         String lon = lonDanger.getText().toString().trim();
         String lat = latDanger.getText().toString().trim();
         String type = typePos + "";
-        String factor = factorDanger.getText().toString().trim();
+        String factor = spFactorType.getSelectedItemPosition() + "";
         String person = personDanger.getText().toString().trim();
         String house = houseDanger.getText().toString().trim();
         String money = moneyDanger.getText().toString().trim();
@@ -666,6 +725,7 @@ public class DangerReportFragment extends Fragment {
         dangerUBean.setMobile(mobile);
         dangerUBean.setName(name);
         dangerUBean.setId(1L);
+        dangerUBean.setDisasterPosition(disasterPosition);
         GreedDaoHelper.insertDanger(dangerUBean);
         List<PhotoPath> queryPhoto = GreedDaoHelper.queryPhoto(2);
         if (queryPhoto != null && queryPhoto.size() > 0) {
@@ -733,13 +793,13 @@ public class DangerReportFragment extends Fragment {
                     tv.setText("正在录音...");
                     File audio = createFileDir("Audio");
                     if (audio != null) {
-                        audioPath = audio.getPath() + "/" + new SimpleDateFormat("yyyyMMddHHmmss").format(new Date()) + ".mp3";
+                        audioPath = audio.getPath() + "/" + new SimpleDateFormat("yyyyMMddHHmmss").format(new Date()) + ".ogg";
                     } else {
                         ToastUtils.showShort("文件夹创建失败");
                     }
                     recorder = new MediaRecorder();
                     recorder.setAudioSource(MediaRecorder.AudioSource.MIC);
-                    recorder.setOutputFormat(MediaRecorder.OutputFormat.DEFAULT);
+                    recorder.setOutputFormat(MediaRecorder.OutputFormat.MPEG_4);
                     recorder.setAudioEncoder(MediaRecorder.AudioEncoder.AAC);
                     recorder.setOutputFile(audioPath);
                     chronometer.start();// 开始计时
@@ -770,12 +830,21 @@ public class DangerReportFragment extends Fragment {
             @Override
             public void onClick(View view) {
                 if (type == 1) {
-                    Uri uri = Uri.parse(tvVideo.getText().toString());
+                    Uri uri;
+                    File videoFile=new File(tvVideo.getText().toString());
+                    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {  //针对Android7.0，需要通过FileProvider封装过的路径，提供给外部调用
+                        uri = FileProvider.getUriForFile(context, "com.nandi.yngsagp.fileprovider", videoFile);//通过FileProvider创建一个content类型的Uri，进行封装
+                    } else { //7.0以下，如果直接拿到相机返回的intent值，拿到的则是拍照的原图大小，很容易发生OOM，所以我们同样将返回的地址，保存到指定路径，返回到Activity时，去指定路径获取，压缩图片
+                        uri = Uri.fromFile(videoFile);
+                    }
                     Intent intent = new Intent(Intent.ACTION_VIEW);
-                    intent.setDataAndType(uri, "video/mp4");
+                    intent.setFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
+                    intent.setDataAndType(uri, "video/*");
                     startActivity(intent);
-                } else {
+                } else if (type == 2) {
                     playAudio(tvAudio.getText().toString());
+                } else {
+                    openFile(tvFile.getText().toString());
                 }
                 popupWindow.dismiss();
             }
@@ -789,13 +858,15 @@ public class DangerReportFragment extends Fragment {
                     }
                     tvVideo.setText("");
                     videoFile = null;
-                } else {
+                } else if (type == 2) {
                     File file = new File(audioPath);
                     if (file.exists()) {
                         file.delete();
                     }
                     tvAudio.setText("");
                     audioPath = null;
+                } else {
+                    tvFile.setText("");
                 }
                 popupWindow.dismiss();
             }
@@ -814,6 +885,11 @@ public class DangerReportFragment extends Fragment {
                 return false;
             }
         });
+    }
+
+    private void openFile(String s) {
+        Intent intent = AppUtils.openFile(s, context);
+        startActivity(intent);
     }
 
     private void playAudio(String s) {
@@ -990,6 +1066,22 @@ public class DangerReportFragment extends Fragment {
                 case TAKE_VIDEO:
                     tvVideo.setText(videoFile.getAbsolutePath());
                     break;
+                case FILE_SELECT_CODE:
+                    Uri fileUri = data.getData();
+                    if (fileUri != null) {
+                        String path = AppUtils.getPath(context, fileUri);
+                        if (path != null) {
+                            String end = path.substring(path.lastIndexOf(".") + 1, path.length()).toLowerCase(Locale.getDefault());
+                            if ("ppt".equals(end) || "pptx".equals(end) || "xls".equals(end)
+                                    || "xlsx".equals(end) || "doc".equals(end) || "docx".equals(end)
+                                    || "pdf".equals(end) || "txt".equals(end)) {
+                                tvFile.setText(path);
+                            } else {
+                                ToastUtils.showShort("不支持上传此类型文件");
+                            }
+                        }
+                    }
+                    break;
             }
         }
     }
@@ -1016,8 +1108,8 @@ public class DangerReportFragment extends Fragment {
         } else if (TextUtils.isEmpty(lonDanger.getText())) {
             ToastUtils.showShort("请获取坐标信息");
             return false;
-        } else if (TextUtils.isEmpty(factorDanger.getText())) {
-            ToastUtils.showShort("请填写诱发因素");
+        } else if (spFactorType.getSelectedItemPosition() == 0) {
+            ToastUtils.showShort("请选择诱发因素");
             return false;
         }
         return true;
